mspm0-adc: remove dynamic vrsel and cleanup
